UI of to do list fixed

+++ app/src/main/java/com/example/arnab/movieinfoimdb/ToWatchActivity.java
@@ -7,11 +7,16 @@ import android.support.annotation.NonNull;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.support.v7.widget.Toolbar;
 import android.util.Log;
+import android.view.Menu;
+import android.view.MenuInflater;
+import android.view.MenuItem;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.ListView;
+import android.widget.TextView;
 import android.widget.Toast;
 
 import com.google.android.gms.tasks.OnCompleteListener;
@@ -42,10 +47,18 @@ public class ToWatchActivity extends AppCompatActivity {
     ArrayList<String> MovieNames = new ArrayList<String>() ;
     //ArrayList<String> MovieInfo = new ArrayList<String>();
     Map<String,Object> MovieInfo = new HashMap<>();
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_to_watch);
+
+        //setting up the toolbar
+        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
+        setSupportActionBar(toolbar);
+        TextView toolbar_title = findViewById(R.id.toolbar_title);
+        toolbar_title.setText("To-Watch");
+        getSupportActionBar().setDisplayShowTitleEnabled(false);
 
         firebaseauthenticator = FirebaseAuth.getInstance();
         user = firebaseauthenticator.getCurrentUser();
@@ -104,48 +117,58 @@ public class ToWatchActivity extends AppCompatActivity {
                     }
 
                 });
-
-
-
-
-
     }
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        MenuInflater menuInflater = getMenuInflater();
+        menuInflater.inflate(R.menu.toolbar_menu_to_do,menu);
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        int id = item.getItemId();
+
+        //noinspection SimplifiableIfStatement
+        if (id == R.id.del_button) {
+            if (name == "") {
+            } else {
+                db.collection("UserToWatchList").document(user_id).collection("ToWatch").document(name)
+                        .delete()
+                        .addOnSuccessListener(new OnSuccessListener<Void>() {
+                            @Override
+                            public void onSuccess(Void aVoid) {
+
+                                Toast.makeText(ToWatchActivity.this, "deleted!", Toast.LENGTH_SHORT).show();
+
+                            }
+                        })
+                        .addOnFailureListener(new OnFailureListener() {
+                            @Override
+                            public void onFailure(@NonNull Exception e) {
+                                Toast.makeText(ToWatchActivity.this, "Unsuccessful!", Toast.LENGTH_SHORT).show();
+                            }
+                        });
+                Intent intent = getIntent();
+                finish();
+                startActivity(intent);
+            }
+        }
+        else if(id == R.id.exit){
+            finish();
+            return true;
+        }
+        else if(id == R.id.add_button){
+            finish();
+            startActivity(new Intent(getApplicationContext(),SearchActivity.class));
+            return true;
+        }
+        return super.onOptionsItemSelected(item);
+    }
+
     private void highlightCurrentRow(View rowView) {
         rowView.setBackgroundColor(Color.GRAY);
-    }
-    public void addMethod(View view){
-        startActivity(new Intent(getApplicationContext(),SearchActivity.class));
-    }
-
-    public void removeMethod(View view){
-
-        if(name == ""){
-            return;
-        }
-
-
-        db.collection("UserToWatchList").document(user_id).collection("ToWatch").document(name)
-                .delete()
-                .addOnSuccessListener(new OnSuccessListener<Void>() {
-                    @Override
-                    public void onSuccess(Void aVoid) {
-
-                        Toast.makeText(ToWatchActivity.this, "deleted!", Toast.LENGTH_SHORT).show();
-
-                    }
-                })
-                .addOnFailureListener(new OnFailureListener() {
-                    @Override
-                    public void onFailure(@NonNull Exception e) {
-                        Toast.makeText(ToWatchActivity.this, "Unsuccessful!", Toast.LENGTH_SHORT).show();
-                    }
-                });
-
-
-        Intent intent = getIntent();
-        finish();
-        startActivity(intent);
-
     }
 
 }
